Revert "security part"
This reverts commit 55cd60e1e3bb1c3c8d8e31f3392b8cf81358a19d.

+++ src/main/java/com/instantcoffee/tech/repo/UserRepo.java
@@ -1,7 +1,8 @@
 package com.instantcoffee.tech.repo;
 
-import com.instantcoffee.tech.authentication.User;
+import com.instantcoffee.tech.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface UserRepo extends JpaRepository<User,Long> {
+public interface UserRepo extends JpaRepository<User,Integer> {
 }
